% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_two_numbers}
\alias{add_two_numbers}
\title{A temp function}
\usage{
add_two_numbers(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{Another number}
}
\value{
Their sum
}
\description{
Adding some meaningless text to test rendering via PR command.
Adding some more
}
