% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusions.R
\name{apply_exclusions}
\alias{apply_exclusions}
\title{Convert case counts in matching rows to NA}
\usage{
apply_exclusions(cases, exclusions)
}
\arguments{
\item{cases}{A dataframe returned by \code{\link[=read_data]{read_data()}}}

\item{exclusions}{A dataframe returned by \code{\link[=read_exclusions]{read_exclusions()}}}
}
\value{
A dataframe with the same rows and schema as \code{cases} and the value in
the column \code{confirm} converted to NA in any rows that match a row in
\code{exclusions}
}
\description{
NAs are skipped in model fitting by EpiNow2, so matched rows are excluded
from the model likelihood.
}
