% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{check_returned_pmf}
\alias{check_returned_pmf}
\title{Run validity checks on the PMF returned from the file}
\usage{
check_returned_pmf(
  pmf_df,
  parameter,
  disease,
  as_of_date,
  geo_value,
  report_date,
  path
)
}
\arguments{
\item{pmf_df}{A dataframe with columns \code{value} and \code{reference_date}.}

\item{parameter}{One of "generation interval", "delay", or "right-truncation"}

\item{disease}{A string specifying the disease being modeled. One of
\code{"COVID-19"} or \code{"Influenza"}.}

\item{as_of_date}{Use the parameters that were used in production on this
date. Set for the current date for the most up-to-to date version of the
parameters and set to an earlier date to use parameters from an earlier
time period.}

\item{geo_value}{An uppercase, two-character string specifying the geographic
value, usually a state or \code{"US"} for national data.}

\item{report_date}{An optional parameter to subset the query to a parameter
on or before a particular \code{report_date}. Right now, the only parameter with
report date-specific estimates is \code{right_truncation}. Note that this
is similar to, but different from \code{as_of_date}. The \code{report_date} is used
to select the particular value of a time-varying estimate. This estimate
may itself be regenerated over time (e.g., as new data becomes available or
with a methodological update). We can pull the estimate for date
\code{report_date} as generated on date \code{as_of_date}.}

\item{path}{A path to a local file}
}
\value{
The unpacked \code{value} column, which is a valid PMF
}
\description{
We're treating this input as possibly invalid because it's from an
external file. We're still updating the schema and this process has
been a frequent source of problems. We want to be alert to any
unexpected changes in schema or format.
}
\seealso{
Other parameters: 
\code{\link{opts_formatter}},
\code{\link{read_disease_parameters}()},
\code{\link{read_interval_pmf}()}
}
\concept{parameters}
