% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{Data}
\alias{Data}
\title{Data Class}
\usage{
Data(
  path = character(0),
  blob_storage_container = character(0),
  report_date = list(),
  reference_date = list(),
  production_date = list()
)
}
\arguments{
\item{path}{A string specifying the path to the data Parquet file.}

\item{blob_storage_container}{Optional. The name of the blob storage
container to which the data file will be uploaded. If NULL, no upload will
occur.}

\item{report_date}{A list of strings representing report dates.}

\item{reference_date}{A list of strings representing reference dates.}

\item{production_date}{A list of strings representing production dates.}
}
\description{
Represents the data-related configurations.
}
