% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{DelayInterval}
\alias{DelayInterval}
\title{DelayInterval Class}
\usage{
DelayInterval(path = character(0), blob_storage_container = character(0))
}
\arguments{
\item{path}{A string specifying the path to the delay interval CSV file.}

\item{blob_storage_container}{Optional. The name of the blob storage
container to get it from. If NULL, will look locally.}
}
\description{
Represents the delay interval parameters.
}
