% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{download_if_specified}
\alias{download_if_specified}
\title{Download if specified}
\usage{
download_if_specified(blob_path, blob_storage_container, dir)
}
\arguments{
\item{blob_path}{The name of the blob to download}

\item{blob_storage_container}{The name of the container to donwload from}

\item{dir}{The directory to which to write the downloaded file}
}
\value{
The path of the file
}
\description{
Download if specified
}
\seealso{
Other azure: 
\code{\link{download_file_from_container}()},
\code{\link{fetch_blob_container}()},
\code{\link{fetch_credential_from_env_var}()}
}
\concept{azure}
