% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{empty_str_if_non_existant}
\alias{empty_str_if_non_existant}
\title{If \code{x} is null or empty, return an empty string, otherwise \code{x}}
\usage{
empty_str_if_non_existant(x)
}
\description{
If \code{x} is null or empty, return an empty string, otherwise \code{x}
}
