% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_diagnostics.R
\name{extract_diagnostics}
\alias{extract_diagnostics}
\title{Extract diagnostic metrics from model fit and data}
\usage{
extract_diagnostics(fit, data, job_id, task_id, disease, geo_value, model)
}
\arguments{
\item{fit}{A list containing the model fit object from \code{epinow2}, which
includes \code{estimates$fit}.}

\item{data}{A data frame containing the input data used in the model fit.}

\item{job_id}{A unique identifier for the job or task being processed.}

\item{task_id}{A unique identifier for the task being performed.}

\item{disease, geo_value, model}{Metadata for downstream processing.}
}
\value{
A \code{data.frame} containing the extracted diagnostic metrics. The
data frame includes the following columns:
\itemize{
\item \code{diagnostic}: The name of the diagnostic metric.
\item \code{value}: The value of the diagnostic metric.
\item \code{state}: The state for which the model was run.
\item \code{disease}: The disease/pathogen being analyzed.
\item \code{job_id}: The unique identifier for the job.
\item \code{task_id}: The unique identifier for the task.
\item \code{disease,geo_value,model}: Metadata for downstream processing.
}
}
\description{
This function extracts various diagnostic metrics from a fitted \code{epinow2}
model and provided data. It checks for low case counts and computes
diagnostics from the fitted model, including the mean acceptance
statistic, divergent transitions, maximum tree depth, and Rhat values.
These diagnostics are then flagged if they exceed specific thresholds,
and the results are returned as a data frame.
}
\details{
The following diagnostics are calculated:
\itemize{
\item \code{mean_accept_stat}: The average acceptance statistic across
all chains.
\item \code{p_divergent}: The proportion of divergent transitions across
all samples.
\item \code{p_max_treedepth}: The proportion of samples that hit the
maximum tree depth.
\item \code{p_high_rhat}: The proportion of parameters with Rhat values
greater than 1.05, indicating potential convergence issues.
\item \code{low_case_count_flag}: A flag indicating if there are low case
counts in the data. See \code{low_case_count_diagnostic()} for more
information on this diagnostic.
\item \code{epinow2_diagnostic_flag}: A combined flag that indicates if
any diagnostic thresholds are exceeded.
}
}
