% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{extract_samples_from_fit}
\alias{extract_samples_from_fit}
\title{Extract and reformat samples from a model fit object}
\usage{
extract_samples_from_fit(fit)
}
\arguments{
\item{fit}{A list representing the model fit object, which includes estimates
and Stan samples. The \code{fit} object must contain an \code{estimates$samples}
data.table with columns \code{date}, \code{time}, \code{parameter}, and \code{type}, as well as a
Stan fit object in \code{estimates$fit}.}
}
\value{
A \code{data.table} in long format with the following columns:
\describe{
\item{\code{time}}{The timepoint at which the draw was made.}
\item{\code{_draw}}{The draw number from the Stan fit.}
\item{\code{_chain}}{The chain number from the Stan fit.}
\item{\code{parameter}}{The standardized parameter name (e.g.,
latent_cases, obs_cases, Rt, growth_rate).}
\item{\code{value}}{The value of the parameter for the corresponding draw
and time.}
\item{\code{reference_date}}{The date associated with the timepoint.}
\item{\code{type}}{The type of the parameter (e.g., estimate or
observation).}
}
}
\description{
This function extracts the desired parameters from a model fit object,
processes the Stan draws, and joins them with a date-time-parameter map for
further analysis. The function transforms the data into long format and
standardizes parameter names for easier quantile summarization and analysis.
}
\details{
The function performs the following steps:
\enumerate{
\item Extracts unique combinations of date, time, parameter, and type from
the model fit.
\item Extracts relevant parameters from the Stan model, such as imputed and
observed reports, the reproduction number (\code{Rt}), and the growth rate
(\code{r}).
\item Converts the extracted data from wide to long format for further
analysis.
\item Performs a left join of the date-time-parameter map onto the Stan
draws to associate the proper dates and types with the parameters.
\item Standardizes the parameter names for consistency.
}
}
