% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{fetch_blob_container}
\alias{fetch_blob_container}
\title{Load Azure Blob endpoint using credentials in environment variables}
\usage{
fetch_blob_container(container_name)
}
\arguments{
\item{container_name}{The Azure Blob Storage container associated with the
credentials}
}
\value{
A Blob endpoint
}
\description{
This \strong{impure} function depends on the environment variables:
\itemize{
\item az_tenant_id
\item az_subscription_id
\item az_resource_group
\item az_storage_account
}
}
\details{
It will error out if any of the above is not set.
}
\seealso{
Other azure: 
\code{\link{download_from_azure_blob}()},
\code{\link{fetch_credential_from_env_var}()}
}
\concept{azure}
