% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{fetch_parameters}
\alias{fetch_parameters}
\title{Load the prod parameters subset to required parameters}
\usage{
fetch_parameters(
  path,
  blob_storage_container,
  local_dest,
  as_of_date,
  disease,
  geo_value
)
}
\arguments{
\item{path}{The path to the local file or in Blob storage}

\item{blob_storage_container}{Either a character with the Azure Blob storage
container with the parameters or the path to the local file}

\item{local_dest}{Only required when \code{blob_storage_container} is specified, a
local directory to save the downloaded file}

\item{as_of_date}{The date that the parameters were used}

\item{disease}{The disease parameters to use}

\item{geo_value}{If a \code{geo_value}-specific parameter, which \code{geo_value} to
use}
}
\value{
A dataframe with columns: "geo_value", "value", "start_date",
"end_date", "parameter", "disease", "format", "id". The "value" column is a
list-column with the PMF of the parameters to be used with EpiNow2 as
non-parametric distribution specification.
}
\description{
Assumes that the dataset is a single parquet file, either sitting locally on
disk or (optionally) sitting in Azure Blob. The flow here is (1) To check if
loading the file from Blob is required (if \code{blob_storage_container} was
specified in the config) (2) Fetch from Blob if needed (3) Load the required
parameters from the downloaded file (4) Validate that the parameters match
expectations (one each of \code{delay}, \code{right_truncation}, and
\code{generation_interval}), (5) Alert success and return the df
}
\details{
The load from disk uses \link{DuckDB}, to enable portability. We can use the same
simple SQL query across languages to get the necessary parameters, and this
can be refactored out into its repo as a future issue.
}
