% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit an EpiNow2 model}
\usage{
fit_model(data, parameters, seed, horizon, priors, sampler_opts)
}
\arguments{
\item{data, }{in the format returned by \code{\link[=read_data]{read_data()}}}

\item{parameters}{As returned from
\code{\link[=read_disease_parameters]{read_disease_parameters()}}}

\item{seed}{The random seed, used for both initialization by EpiNow2 in R and
sampling in Stan}

\item{horizon}{The number of days, as an integer, to forecast}

\item{priors}{A list of lists. The first level should contain the key \code{rt}
with elements \code{mean} and \code{sd} and the key \code{gp} with element \code{alpha_sd}.}

\item{sampler_opts}{A list. The Stan sampler options to be passed through
EpiNow2. It has required keys: \code{cores}, \code{chains}, \code{iter_warmup},
\code{iter_sampling}, \code{max_treedepth}, and \code{adapt_delta}.}
}
\value{
A fitted model object of class \code{epinow} or, if model fitting fails,
an NA is returned with a warning
}
\description{
Fit an EpiNow2 model
}
\seealso{
Other pipeline: 
\code{\link{orchestrate_pipeline}()}
}
\concept{pipeline}
