% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{GenerationInterval}
\alias{GenerationInterval}
\title{GenerationInterval Class}
\usage{
GenerationInterval(path = character(0), blob_storage_container = character(0))
}
\arguments{
\item{path}{A string specifying the path to the generation interval CSV file.}

\item{blob_storage_container}{Optional. The name of the blob storage
container to get it from. If NULL, will look locally.}
}
\description{
Represents the generation interval parameters.
}
