% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{GpPrior}
\alias{GpPrior}
\title{GpPrior Class}
\usage{
GpPrior(alpha_sd = integer(0))
}
\arguments{
\item{alpha_sd}{A numeric value representing the standard deviation of the
alpha parameter in the GP prior.}
}
\description{
Represents the Gaussian Process prior parameters.
}
