% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{load_parameters}
\alias{load_parameters}
\title{Load the parameters from a local file}
\usage{
load_parameters(path, disease, geo_value, as_of_date)
}
\arguments{
\item{path}{Location of the local file on disk}

\item{disease}{As specified in the prod parameter dataset}

\item{geo_value}{As specified in tbe prod parameter dataset}

\item{as_of_date}{The parameters used as of a particular date. Allows using
an older version of the parameters, such as for backtesting.}
}
\value{
A dataframe with columns: "geo_value", "value", "start_date",
"end_date", "parameter", "disease", "format", "id"
}
\description{
Load the parameters from a local file
}
