% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_low_count_threshold.R
\name{low_case_count_threshold}
\alias{low_case_count_threshold}
\title{Determine Low Case Count Threshold Based on Pathogen}
\usage{
low_case_count_threshold(low_case_count_thresholds, disease)
}
\arguments{
\item{low_case_count_thresholds}{A list of thresholds to use for
determining n_low_case_count in diagnostic file}

\item{disease}{A string specifying the disease being modeled. One of
\code{"COVID-19"} or \code{"Influenza"} or \code{"RSV"}.}
}
\value{
An integer that reflects the value X where number ED visits < X
in the past week and week prior results in an n_low_case_count flag for
that pathogen-state pair
}
\description{
Determine Low Case Count Threshold Based on Pathogen
}
\seealso{
Other diagnostics: 
\code{\link{extract_diagnostics}()},
\code{\link{low_case_count_diagnostic}()}
}
\concept{diagnostics}
