% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{Parameters}
\alias{Parameters}
\title{Parameters Class}
\usage{
Parameters(
  generation_interval = GenerationInterval(),
  delay_interval = DelayInterval(),
  right_truncation = RightTruncation()
)
}
\arguments{
\item{generation_interval}{An instance of \code{GenerationInterval} class.}

\item{delay_interval}{An instance of \code{DelayInterval} class.}

\item{right_truncation}{An instance of \code{RightTruncation} class.}
}
\description{
Holds all parameter-related configurations for the pipeline.
}
