% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{Priors}
\alias{Priors}
\title{Priors Class}
\usage{
Priors(rt = RtPrior(), gp = GpPrior())
}
\arguments{
\item{rt}{An instance of \code{RtPrior} class.}

\item{gp}{An instance of \code{GpPrior} class.}
}
\description{
Holds all prior-related configurations for the pipeline.
}
