% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{process_pipeline}
\alias{process_pipeline}
\title{Run model fitting process}
\usage{
process_pipeline(config, output_dir)
}
\value{
TRUE for pipeline run success or throws an error as a side-effect
for failure of a step.
}
\description{
This part of the pipeline contains the meat of the logic. The logic in this
piece of the pipeline run normally,
}
