% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{read_disease_parameters}
\alias{read_disease_parameters}
\title{Read in disease process parameters from an external file or files}
\usage{
read_disease_parameters(
  generation_interval_path,
  delay_interval_path,
  right_truncation_path,
  disease,
  as_of_date,
  geo_value,
  report_date
)
}
\arguments{
\item{generation_interval_path, delay_interval_path, right_truncation_path}{Path to a local file with the parameter PMF. See \link{read_interval_pmf} for
details on the file schema. The parameters can be in the same file or a
different file.}

\item{disease}{A string specifying the disease being modeled. One of
\code{"COVID-19"} or \code{"Influenza"} or \code{"RSV"}.}

\item{as_of_date}{Use the parameters that were used in production on this
date. Set for the current date for the most up-to-to date version of the
parameters and set to an earlier date to use parameters from an earlier
time period.}

\item{geo_value}{An uppercase, two-character string specifying the geographic
value, usually a state or \code{"US"} for national data.}

\item{report_date}{An optional parameter to subset the query to a parameter
on or before a particular \code{report_date}. Right now, the only parameter with
report date-specific estimates is \code{right_truncation}. Note that this
is similar to, but different from \code{as_of_date}. The \code{report_date} is used
to select the particular value of a time-varying estimate. This estimate
may itself be regenerated over time (e.g., as new data becomes available or
with a methodological update). We can pull the estimate for date
\code{report_date} as generated on date \code{as_of_date}.}
}
\value{
A named list with three PMFs. The list elements are named
\code{generation_interval}, \code{delay_interval}, and \code{right_truncation}. If a path
to a local file is not provided (NA or NULL), the corresponding parameter
estimate will be NA in the returned list.
}
\description{
Read in disease process parameters from an external file or files
}
\details{
\code{generation_interval_path} is required because the generation
interval is a required parameter for $R_t$ estimation.
\code{delay_interval_path} and \code{right_truncation_path} are optional (but
strongly suggested).
}
\seealso{
Other parameters: 
\code{\link{opts_formatter}},
\code{\link{read_interval_pmf}()}
}
\concept{parameters}
