% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{read_interval_pmf}
\alias{read_interval_pmf}
\title{Read parameter PMF into memory}
\usage{
read_interval_pmf(
  path,
  disease = c("COVID-19", "Influenza", "test"),
  as_of_date,
  parameter = c("generation_interval", "delay", "right_truncation"),
  group = NA
)
}
\arguments{
\item{path}{A path to a local file}

\item{disease}{One of "COVID-19" or "Influenza"}

\item{as_of_date}{The parameters "as of" the date of the model run}

\item{parameter}{One of "generation interval", "delay", or "right-truncation}

\item{group}{An optional parameter to subset the query to a parameter with a
particular two-letter state abbrevation. Right now, the only parameter with
state-specific estimates is \code{right-truncation}.}
}
\value{
A PMF vector
}
\description{
Using DuckDB from a parquet file. The function expects the file to be in SCD2
format with column names:
\itemize{
\item parameter
\item geo_value
\item disease
\item start_date
\item end_date
\item value
}
}
\details{
start_date and end_date specify the date range for which the value was used.
end_date may be NULL (e.g. for the current value used in production). value
must contain a pmf vector whose values are all positive and sum to 1. all
other fields must be consistent with the specifications of the function
arguments describe below, which are used to query from the .parquet file.

SCD2 format is shorthand for slowly changing dimension type 2. This format is
normalized to track change over time:
https://en.wikipedia.org/wiki/Slowly_changing_dimension#Type_2:_add_new_row
}
