% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_json_into_config}
\alias{read_json_into_config}
\title{Read JSON Configuration into Config Object}
\usage{
read_json_into_config(config_path)
}
\arguments{
\item{config_path}{A string specifying the path to the JSON configuration
file.}
}
\value{
An instance of the \code{Config} class populated with the data from the
JSON file.
}
\description{
Reads a JSON file from the specified path and converts it into a \code{Config}
object.
}
