% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{read_parameters}
\alias{read_parameters}
\title{Read in disease process parameters from an external file or files}
\usage{
read_parameters(
  generation_interval_path,
  delay_interval_path,
  right_truncation_path,
  disease,
  as_of_date,
  group
)
}
\arguments{
\item{generation_interval_path, delay_interval_path, right_truncation_path}{Path to a local file with the parameter PMF. See \link{read_interval_pmf} for
details on the file schema. The parameters can be in the same file or a
different file.}

\item{disease}{One of \code{COVID-19} or \code{Influenza}}

\item{as_of_date}{Use the parameters that were used in production on this
date. Set for the current date for the most up-to-to date version of the
parameters and set to an earlier date to use parameters from an earlier
time period.}

\item{group}{Used only for parameters with a state-level estimate (i.e., only
right-truncation). The two-letter uppercase state abbreviation.}
}
\value{
A named list with three PMFs. The list elements are named
\code{generation_interval}, \code{delay_interval}, and \code{right_truncation}. If a path
to a local file is not provided (NA or NULL), the corresponding parameter
estimate will be NA in the returned list.
}
\description{
Read in disease process parameters from an external file or files
}
\details{
\code{generation_interval_path} is required because the generation
interval is a required parameter for $R_t$ estimation.
\code{delay_interval_path} and \code{right_truncation_path} are optional (but
strongly suggested).
}
