% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{RightTruncation}
\alias{RightTruncation}
\title{RightTruncation Class}
\usage{
RightTruncation(path = character(0), blob_storage_container = character(0))
}
\arguments{
\item{path}{A string specifying the path to the right truncation CSV file.}

\item{blob_storage_container}{Optional. The name of the blob storage
container to get it from. If NULL, will look locally.}
}
\description{
Represents the right truncation parameters.
}
