% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{RtPrior}
\alias{RtPrior}
\title{RtPrior Class}
\usage{
RtPrior(mean = integer(0), sd = integer(0))
}
\arguments{
\item{mean}{A numeric value representing the mean of the Rt prior.}

\item{sd}{A numeric value representing the standard deviation of the Rt
prior.}
}
\description{
Represents the Rt prior parameters.
}
