% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{sample_processing_functions}
\alias{sample_processing_functions}
\alias{process_samples}
\alias{process_quantiles}
\title{Process posterior samples from a Stan fit object (raw draws).}
\usage{
process_samples(fit, geo_value, model, disease)

process_quantiles(fit, geo_value, model, disease)
}
\arguments{
\item{fit}{An EpiNow2 fit object with posterior estimates.}

\item{disease, geo_value, model}{Metadata for downstream processing.}
}
\value{
A data.table of posterior draws or quantiles, merged and processed.
}
\description{
Extracts raw posterior samples from a Stan fit object and post-processes
them, including merging with a fact table and standardizing the parameter
names. If calling \code{process_quantiles()} the 50\% and 95\% intervals are
returned in \code{{tidybayes}} format.
}
