% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{SamplerOpts}
\alias{SamplerOpts}
\title{SamplerOpts Class}
\usage{
SamplerOpts(
  cores = integer(0),
  chains = integer(0),
  iter_warmup = integer(0),
  iter_sampling = integer(0),
  adapt_delta = integer(0),
  max_treedepth = integer(0)
)
}
\arguments{
\item{cores}{An integer specifying the number of CPU cores to use.}

\item{chains}{An integer specifying the number of Markov chains.}

\item{iter_warmup}{An integer specifying the number of warmup iterations.}

\item{iter_sampling}{An integer specifying the number of sampling iterations.}

\item{adapt_delta}{A numeric value for the target acceptance probability.}

\item{max_treedepth}{An integer specifying the maximum tree depth for the
sampler.}
}
\description{
Represents the sampler options for the pipeline.
}
