% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_config}
\alias{validate_config}
\title{Compare loaded JSON against expectation in \code{inst/data/config-schema.json}}
\usage{
validate_config(
  config_path,
  config_schema_path = system.file("extdata/config_schema.json", package =
    "CFAEpiNow2Pipeline")
)
}
\arguments{
\item{config_path}{The path to the config file, either in the local
filesystem or with an Azure Blob Storage container. If
\code{blob_storage_container} is specified, the path is assumed to be within
the specified container otherwise it is assumed to be in the local
filesystem.}

\item{config_schema_path}{The path to the file holding the schema for the
config json for the validator to use.}
}
\value{
NULL, invisibly
}
\description{
Compare loaded JSON against expectation in \code{inst/data/config-schema.json}
}
