% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{validate_parameters}
\alias{validate_parameters}
\title{Check loaded parameters df matches expected}
\usage{
validate_parameters(df)
}
\arguments{
\item{df}{The parameters df returned by \code{\link[=load_parameters]{load_parameters()}}}
}
\value{
The argument df, invisibly
}
\description{
We want one each of \code{right_truncation}, \code{delay}, and \code{generation_interval}
}
