% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{Config}
\alias{Config}
\title{Config Class}
\usage{
Config(
  job_id = character(0),
  task_id = character(0),
  min_reference_date = (function (.data = double()) 
 {
     .Date(.data)
 })(),
  max_reference_date = (function (.data = double()) 
 {
     .Date(.data)
 })(),
  report_date = (function (.data = double()) 
 {
     .Date(.data)
 })(),
  disease = character(0),
  geo_value = character(0),
  geo_type = character(0),
  data = Data(),
  seed = integer(0),
  horizon = integer(0),
  priors = Priors(),
  parameters = Parameters(),
  sampler_opts = SamplerOpts(),
  exclusions = Exclusions(),
  config_version = integer(0),
  quantile_width = list(),
  model = character(0)
)
}
\arguments{
\item{job_id}{A string specifying the job.}

\item{task_id}{A string specifying the task.}

\item{min_reference_date}{A Date object representing the minimum reference
date.}

\item{max_reference_date}{A Date object representing the maximum reference
date.}

\item{report_date}{A Date object representing the report date.}

\item{disease}{A string specifying the disease being modeled.}

\item{geo_value}{A string specifying the geographic value, usually a state.}

\item{geo_type}{A string specifying the geographic type, usually "state".}

\item{data}{An instance of \code{Data} class containing data configurations.}

\item{seed}{An integer for setting the random seed.}

\item{horizon}{An integer specifying the forecasting horizon.}

\item{priors}{An instance of \code{Priors} class containing prior configurations.}

\item{parameters}{An instance of \code{Parameters} class containing parameter
configurations.}

\item{sampler_opts}{An instance of \code{SamplerOpts} class containing sampler
options.}

\item{exclusions}{An instance of \code{Exclusions} class containing exclusion
criteria.}

\item{config_version}{A numeric value specifying the configuration version.}

\item{quantile_width}{A list of numeric values representing the desired
quantiles.}

\item{model}{A string specifying the model to be used.}
}
\description{
Represents the complete configuration for the pipeline.
}
