% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{Config}
\alias{Config}
\title{Config Class}
\usage{
Config(
  job_id = character(0),
  task_id = character(0),
  min_reference_date = character(0),
  max_reference_date = character(0),
  report_date = character(0),
  disease = character(0),
  geo_value = character(0),
  geo_type = character(0),
  seed = integer(0),
  horizon = integer(0),
  model = "EpiNow2",
  config_version = character(0),
  quantile_width = c(0.5, 0.95),
  data = Data(),
  priors = list(),
  parameters = Parameters(),
  sampler_opts = list(),
  exclusions = Exclusions()
)
}
\arguments{
\item{job_id}{A string specifying the job.}

\item{task_id}{A string specifying the task.}

\item{min_reference_date}{A string representing the minimum reference
date. Formatted as "YYYY-MM-DD".}

\item{max_reference_date}{A string representing the maximum reference
date. Formatted as "YYYY-MM-DD".}

\item{report_date}{A string representing the report date. Formatted as
"YYYY-MM-DD".}

\item{disease}{A string specifying the disease being modeled.}

\item{geo_value}{A string specifying the geographic value, usually a state.}

\item{geo_type}{A string specifying the geographic type, usually "state".}

\item{seed}{An integer for setting the random seed.}

\item{horizon}{An integer specifying the forecasting horizon.}

\item{model}{A string specifying the model to be used.}

\item{config_version}{A numeric value specifying the configuration version.}

\item{quantile_width}{A vector of numeric values representing the desired
quantiles.}

\item{data}{An instance of \code{Data} class containing data configurations.}

\item{priors}{A list of lists. The first level should contain the key \code{rt}
with elements \code{mean} and \code{sd} and the key \code{gp} with element \code{alpha_sd}.}

\item{parameters}{An instance of \code{Parameters} class containing parameter
configurations.}

\item{sampler_opts}{A list. The Stan sampler options to be passed through
EpiNow2. It has required keys: \code{cores}, \code{chains}, \code{iter_warmup},
\code{iter_sampling}, \code{max_treedepth}, and \code{adapt_delta}.}

\item{exclusions}{An instance of \code{Exclusions} class containing exclusion
criteria.}
}
\description{
Represents the complete configuration for the pipeline.
}
