% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{Config}
\alias{Config}
\title{Config Class}
\usage{
Config(
  job_id = class_missing,
  task_id = class_missing,
  min_reference_date = class_missing,
  max_reference_date = class_missing,
  report_date = class_missing,
  production_date = class_missing,
  disease = class_missing,
  low_case_count_thresholds = class_missing,
  geo_value = class_missing,
  geo_type = class_missing,
  seed = class_missing,
  horizon = class_missing,
  model = class_missing,
  config_version = class_missing,
  quantile_width = class_missing,
  data = class_missing,
  priors = class_missing,
  parameters = class_missing,
  sampler_opts = class_missing,
  exclusions = class_missing,
  output_container = class_missing
)
}
\arguments{
\item{job_id}{A string specifying the job.}

\item{task_id}{A string specifying the task.}

\item{min_reference_date}{A string representing the minimum reference
date. Formatted as "YYYY-MM-DD".}

\item{max_reference_date}{A string representing the maximum reference
date. Formatted as "YYYY-MM-DD".}

\item{report_date}{A string representing the report date. Formatted as
"YYYY-MM-DD".}

\item{production_date}{A string representing the production date.
Formatted as "YYYY-MM-DD".}

\item{disease}{A string specifying the disease being modeled. One of
\code{"COVID-19"} or \code{"Influenza"} or \code{"RSV"}.}

\item{low_case_count_thresholds}{A named list of thresholds to use for
determining n_low_case_count in diagnostic file
Example: list(\code{COVID-19} = 10, \verb{Influenza`` = 10, }RSV` = 5)}

\item{geo_value}{An uppercase, two-character string specifying the geographic
value, usually a state or \code{"US"} for national data.}

\item{geo_type}{A string specifying the geographic type, usually "state".}

\item{seed}{An integer for setting the random seed.}

\item{horizon}{An integer specifying the forecasting horizon.}

\item{model}{A string specifying the model to be used.}

\item{config_version}{A numeric value specifying the configuration version.}

\item{quantile_width}{A vector of numeric values representing the desired
quantiles. Passed to \code{\link[tidybayes:reexports]{tidybayes::median_qi()}}.}

\item{data}{An instance of \code{Data} class containing data configurations.}

\item{priors}{A list of lists. The first level should contain the key \code{rt}
with elements \code{mean} and \code{sd} and the key \code{gp} with element \code{alpha_sd}.}

\item{parameters}{An instance of \code{Parameters} class containing parameter
configurations.}

\item{sampler_opts}{A list. The Stan sampler options to be passed through
EpiNow2. It has required keys: \code{cores}, \code{chains}, \code{iter_warmup},
\code{iter_sampling}, \code{max_treedepth}, and \code{adapt_delta}.}

\item{exclusions}{An instance of \code{Exclusions} class containing exclusion
criteria.}

\item{output_container}{An optional string specifying the output blob storage
container.}
}
\description{
Represents the complete configuration for the pipeline.
}
\seealso{
Other config: 
\code{\link{Data}()},
\code{\link{Interval}},
\code{\link{Parameters}()},
\code{\link{read_json_into_config}()}
}
\concept{config}
