% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{Exclusions}
\alias{Exclusions}
\title{Exclusions Class}
\usage{
Exclusions(path = character(0))
}
\arguments{
\item{path}{A string specifying the path to a CSV file containing exclusion
data. It should include at least the columns: \code{reference_date},
\code{report_date}, ' \code{state_abb}, \code{disease}.}
}
\description{
Represents exclusion criteria for the pipeline.
}
