% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{extract_diagnostics}
\alias{extract_diagnostics}
\title{Extract diagnostic metrics from model fit and data}
\usage{
extract_diagnostics(fit, data, job_id, task_id, disease, geo_value, model)
}
\arguments{
\item{fit}{The model fit object from \code{EpiNow2}}

\item{data}{A data frame containing the input data used in the model fit.}

\item{job_id}{A unique identifier for the job}

\item{task_id}{A unique identifier for the task}

\item{disease, geo_value, model}{Metadata for downstream processing.}
}
\value{
A \code{data.frame} containing the extracted diagnostic metrics. The
data frame includes the following columns:
\itemize{
\item \code{diagnostic}: The name of the diagnostic metric.
\item \code{value}: The value of the diagnostic metric.
\item \code{job_id}: The unique identifier for the job.
\item \code{task_id}: The unique identifier for the task.
\item \code{disease,geo_value,model}: Metadata for downstream processing.
}
}
\description{
This function extracts various diagnostic metrics from a fitted \code{EpiNow2}
model and provided data. It checks for low case counts and computes
diagnostics from the fitted model, including the mean acceptance
statistic, divergent transitions, maximum tree depth, and Rhat values.
Additionally, a combined flag is computed indicating if any diagnostics
are outside an acceptable range. The results are returned as a data frame.
}
\details{
The following diagnostics are calculated:
\itemize{
\item \code{mean_accept_stat}: The average acceptance statistic across
all chains.
\item \code{p_divergent}: The \emph{proportion} of divergent transitions across
all samples.
\item \code{n_divergent}: The \emph{number} of divergent transitions across
all samples.
\item \code{p_max_treedepth}: The proportion of samples that hit the
maximum tree depth.
\item \code{p_high_rhat}: The \emph{proportion} of parameters with Rhat values
greater than 1.05, indicating potential convergence issues.
\item \code{n_high_rhat}: The \emph{number} of parameters with Rhat values
greater than 1.05, indicating potential convergence issues.
\item \code{low_case_count_flag}: A flag indicating if there are low case
counts in the data. See \code{low_case_count_diagnostic()} for more
information on this diagnostic.
\item \code{epinow2_diagnostic_flag}: A combined flag that indicates if
any diagnostic metrics are outside an accepted range, as determined
by the thresholds: (1) mean_accept_stat < 0.1, (2) p_divergent >
0.0075, (3) p_max_treedepth > 0.05, and (4) p_high_rhat > 0.0075.
}
}
\seealso{
Other diagnostics: 
\code{\link{low_case_count_diagnostic}()}
}
\concept{diagnostics}
