% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{fetch_blob_container}
\alias{fetch_blob_container}
\title{Load Azure Blob endpoint using credentials in environment variables}
\usage{
fetch_blob_container(container_name)
}
\arguments{
\item{container_name}{The Azure Blob Storage container associated with the
credentials}
}
\value{
A Blob endpoint from \code{\link[AzureStor:storage_endpoint]{AzureStor::blob_endpoint()}}
}
\description{
This \strong{impure} function depends on the environment variables:
\itemize{
\item TENANT_ID
\item SUBSCRIPTION
\item RESOURCE_GROUP
\item STORAGE_ACCOUNT
}
}
\details{
It will error out if any of the above is not set.
}
