% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{fetch_credential_from_env_var}
\alias{fetch_credential_from_env_var}
\title{Fetch Azure credential from environment variable}
\usage{
fetch_credential_from_env_var(env_var)
}
\arguments{
\item{env_var}{A character, the credential to fetch}
}
\value{
The associated value
}
\description{
And throw an informative error if credential is not found
}
\seealso{
Other azure: 
\code{\link{download_file_from_container}()},
\code{\link{download_if_specified}()},
\code{\link{fetch_blob_container}()}
}
\concept{azure}
