% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{format_stan_opts}
\alias{format_stan_opts}
\title{Format Stan options for input to EpiNow2}
\usage{
format_stan_opts(sampler_opts, seed)
}
\arguments{
\item{sampler_opts}{A list. The Stan sampler options to be passed through
EpiNow2. It has required keys: \code{cores}, \code{chains}, \code{iter_warmup},
\code{iter_sampling}, \code{max_treedepth}, and \code{adapt_delta}.}

\item{seed}{A stochastic seed passed here to the Stan sampler and as the R
PRNG seed for EpiNow2 initialization}
}
\value{
A \code{stan_opts} object of arguments
}
\description{
Format configuration \code{sampler_opts} for input to EpiNow2 via a call to
\code{\link[EpiNow2:stan_opts]{EpiNow2::stan_opts()}}.
}
\seealso{
Other pipeline: 
\code{\link{fit_model}()},
\code{\link{orchestrate_pipeline}()}
}
\concept{pipeline}
