% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{Interval}
\alias{Interval}
\alias{GenerationInterval}
\alias{DelayInterval}
\alias{RightTruncation}
\title{Interval Class}
\usage{
Interval(path = character(0), blob_storage_container = character(0))

GenerationInterval(path = character(0), blob_storage_container = character(0))

DelayInterval(path = character(0), blob_storage_container = character(0))

RightTruncation(path = character(0), blob_storage_container = character(0))
}
\arguments{
\item{path}{A string specifying the path to the generation interval CSV file.}

\item{blob_storage_container}{Optional. The name of the blob storage
container to get it from. If NULL, will look locally.}
}
\description{
Represents a generic interval. Meant to be subclassed.

Represents the generation interval parameters.

Represents the delay interval parameters.

Represents the right truncation parameters.
}
\seealso{
Other config: 
\code{\link{Config}()},
\code{\link{Exclusions}()},
\code{\link{Parameters}()},
\code{\link{read_json_into_config}()}

Other config: 
\code{\link{Config}()},
\code{\link{Exclusions}()},
\code{\link{Parameters}()},
\code{\link{read_json_into_config}()}

Other config: 
\code{\link{Config}()},
\code{\link{Exclusions}()},
\code{\link{Parameters}()},
\code{\link{read_json_into_config}()}

Other config: 
\code{\link{Config}()},
\code{\link{Exclusions}()},
\code{\link{Parameters}()},
\code{\link{read_json_into_config}()}
}
\concept{config}
