% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_diagnostics.R
\name{low_case_count_diagnostic}
\alias{low_case_count_diagnostic}
\title{Calculate low case count diagnostic flag}
\usage{
low_case_count_diagnostic(df)
}
\arguments{
\item{df}{A dataframe as returned by \code{\link[=read_data]{read_data()}}. The dataframe must
include columns such as \code{reference_date} (a date vector) and \code{confirm}
(the number of confirmed cases per day).}
}
\value{
A logical value (TRUE or FALSE) indicating whether either of the last
two weeks in the dataset had fewer than 10 cases per week.
}
\description{
The diagnostic flag is TRUE if either of the \emph{last} two weeks of the dataset
have fewer than an aggregate 10 cases per week. This aggregation excludes the
count from confirmed outliers, which have been set to NA in the data.
}
\details{
This function assumes that the \code{df} input dataset has been
"completed": that any implicit missingness has been made explicit.
}
