% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{opts_formatter}
\alias{opts_formatter}
\alias{format_generation_interval}
\alias{format_delay_interval}
\alias{format_right_truncation}
\title{Format PMFs for EpiNow2}
\usage{
format_generation_interval(pmf)

format_delay_interval(pmf)

format_right_truncation(pmf, data)
}
\arguments{
\item{pmf}{As returned from \code{\link[=read_disease_parameters]{read_disease_parameters()}}.
A PMF vector or an NA, if not applying the PMF to the model fit.}

\item{data}{As returned from \code{\link[=read_data]{read_data()}}}
}
\value{
An EpiNow2::*_opts() formatted object or NA with a message
}
\description{
Format PMFs for use by EpiNow2. Delays or right truncation are optional and
can be skipped by passing an NA.
}
