% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{opts_formatter}
\alias{opts_formatter}
\alias{format_generation_interval}
\alias{format_delay_interval}
\alias{format_right_truncation}
\title{Format PMFs for EpiNow2}
\usage{
format_generation_interval(pmf)

format_delay_interval(pmf)

format_right_truncation(pmf, data)
}
\arguments{
\item{pmf}{As returned by \code{\link[=read_disease_parameters]{read_disease_parameters()}}. A PMF vector or an NA,
if not applying the PMF to the model fit.}

\item{data}{in the format returned by \code{\link[=read_data]{read_data()}}}
}
\value{
An \verb{EpiNow2::*_opts()} formatted object or NA with a message
}
\description{
Opinionated wrappers around \code{\link[EpiNow2:generation_time_opts]{EpiNow2::generation_time_opts()}},
\code{\link[EpiNow2:delay_opts]{EpiNow2::delay_opts()}}, or \code{\link[EpiNow2:dist_spec]{EpiNow2::dist_spec()}} which format the
generation interval, delay, or right truncation parameters as an object ready
for input to \code{EpiNow2}.
}
\details{
Delays or right truncation are optional and can be skipped by passing \code{pmf = NA}.
}
\seealso{
Other parameters: 
\code{\link{check_returned_pmf}()},
\code{\link{read_disease_parameters}()},
\code{\link{read_interval_pmf}()}
}
\concept{parameters}
