% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read in the dataset of incident case counts}
\usage{
read_data(
  data_path,
  disease = c("COVID-19", "Influenza", "RSV", "test"),
  geo_value,
  report_date,
  max_reference_date,
  min_reference_date,
  facility_active_proportion
)
}
\arguments{
\item{data_path}{The path to the local file. This could contain a glob and
must be in parquet format.}

\item{disease}{A string specifying the disease being modeled. One of
\code{"COVID-19"} or \code{"Influenza"} or \code{"RSV"}.}

\item{geo_value}{An uppercase, two-character string specifying the geographic
value, usually a state or \code{"US"} for national data.}

\item{report_date}{A string representing the report date. Formatted as
"YYYY-MM-DD".}

\item{max_reference_date}{A string representing the maximum reference
date. Formatted as "YYYY-MM-DD".}

\item{min_reference_date}{A string representing the minimum reference
date. Formatted as "YYYY-MM-DD".}

\item{facility_active_proportion}{A numeric value between 0 and 1 specifying
the proportion of days during the modeling period that facilities must have
valid DDI codes to be included in the analysis. Default is 0.94 (facilities
must have visits every day). Lower values allow inclusion of facilities
with fewer active days.}
}
\value{
A dataframe with one or more rows and columns \code{report_date},
\code{reference_date}, \code{geo_value}, \code{confirm}
}
\description{
Each row of the table corresponds to a single facilities' cases for a
reference-date/report-date/disease tuple. We want to aggregate these counts
to the level of geographic aggregate/report-date/reference-date/disease.
}
\details{
We handle two distinct cases for geographic aggregates:
\enumerate{
\item A single state: Subset to facilities \strong{in that state only} and aggregate
up to the state level 2. The US overall: Aggregate over all facilities
without any subsetting
}

Note that we do \emph{not} apply exclusions here. The exclusions are applied
later, after the aggregations. That means that for the US overall, we
aggregate over points that might potentially be excluded at the state level.
Our recourse in this case is to exclude the US overall aggregate point.
}
\concept{read_data}
