% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read in the dataset of incident case counts}
\usage{
read_data(
  data_path,
  disease = c("COVID-19", "Influenza", "RSV", "test"),
  geo_value,
  report_date,
  max_reference_date,
  min_reference_date
)
}
\arguments{
\item{data_path}{The path to the local file. This could contain a glob and
must be in parquet format.}
}
\value{
A dataframe with one or more rows and columns \code{report_date},
\code{reference_date}, \code{geo_value}, \code{confirm}
}
\description{
Each row of the table corresponds to a single facilities' cases for a
reference-date/report-date/disease tuple. We want to aggregate these counts
to the level of geographic aggregate/report-date/reference-date/disease.
}
\details{
We handle two distinct cases for geographic aggregates:
\enumerate{
\item A single state: Subset to facilities \strong{in that state only} and aggregate
up to the state level 2. The US overall: Aggregate over all facilities
without any subsetting
}

Note that we do \emph{not} apply exclusions here. The exclusions are applied
later, after the aggregations. That means that for the US overall, we
aggregate over points that might potentially be excluded at the state level.
Our recourse in this case is to exclude the US overall aggregate point.
}
\concept{read_data}
