% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusions.R
\name{read_exclusions}
\alias{read_exclusions}
\title{Read exclusions from an external file}
\usage{
read_exclusions(path)
}
\arguments{
\item{path}{The path to the exclusions file in \code{.csv} format}
}
\value{
A dataframe with columns \code{reference_date}, \code{report_date},
\code{geo_value}, \code{disease}
}
\description{
Expects to read a CSV with required columns:
\itemize{
\item \code{reference_date}
\item \code{report_date}
\item \code{state}
\item \code{disease}
}
}
\details{
These columns have the same meaning as in \code{\link[=read_data]{read_data()}}. Additional columns
are allowed and will be ignored by the reader.
}
\seealso{
Other exclusions: 
\code{\link{apply_exclusions}()}
}
\concept{exclusions}
