% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{read_parameters}
\alias{read_parameters}
\title{Read in disease process parameters from an external file or files}
\usage{
read_parameters(
  generation_interval_path,
  delay_interval_path,
  right_truncation_path,
  disease,
  as_of_date,
  state
)
}
\arguments{
\item{generation_interval_path, delay_interval_path, right_truncation_path}{Path to a local file with the parameter PMF. See \link{read_interval_pmf} for
details on the file schema. The parameters can be in the same file or a
different file.}

\item{disease}{One of \code{COVID-19} or \code{Influenza}}

\item{as_of_date}{The date "as of" the run. Set for the current date for the
most up-to-to date version of the parameters and set to an earlier date to
use parameters from an earlier time period.}

\item{state}{Used only for right-truncation, the two-letter uppercase state
abbrevation.}
}
\value{
A named list with three PMFs. The list elements are named
\code{generation_interval}, \code{delay_interval}, and \code{right_truncation}. If a path
is
}
\description{
Read in disease process parameters from an external file or files
}
