% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{sample_processing_functions}
\alias{sample_processing_functions}
\alias{process_samples}
\alias{process_quantiles}
\title{Process posterior samples from a Stan fit object (raw draws).}
\usage{
process_samples(fit, geo_value, model, disease)

process_quantiles(fit, geo_value, model, disease, quantile_width)
}
\arguments{
\item{fit}{An \code{EpiNow2} fit object with posterior estimates.}

\item{geo_value}{An uppercase, two-character string specifying the geographic
value, usually a state or \code{"US"} for national data.}

\item{model}{A string specifying the model to be used.}

\item{disease}{A string specifying the disease being modeled. One of
\code{"COVID-19"} or \code{"Influenza"} or \code{"RSV"}.}

\item{quantile_width}{A vector of numeric values representing the desired
quantiles. Passed to \code{\link[tidybayes:reexports]{tidybayes::median_qi()}}.}
}
\value{
A data.table of posterior draws or quantiles, merged and processed.
}
\description{
Extracts raw posterior samples from a Stan fit object and post-processes
them, including merging with a fact table and standardizing the parameter
names. If calling \verb{[process_quantiles()]} the 50\% and 95\% intervals are
returned in \code{tidybayes} format.
}
\seealso{
Other write_output: 
\code{\link{write_model_outputs}()},
\code{\link{write_output_dir_structure}()}
}
\concept{write_output}
