% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{write_model_outputs}
\alias{write_model_outputs}
\title{Write model outputs to specified directories}
\usage{
write_model_outputs(
  fit,
  samples,
  summaries,
  output_dir,
  job_id,
  task_id,
  metadata = list(),
  diagnostics
)
}
\arguments{
\item{fit}{An \code{EpiNow2} fit object with posterior estimates.}

\item{samples}{A data.table as returned by \code{\link[=process_samples]{process_samples()}}}

\item{summaries}{A data.table as returned by \code{\link[=process_quantiles]{process_quantiles()}}}

\item{output_dir}{A string specifying the directory where output, logs, and
other pipeline artifacts will be saved. Defaults to the root directory ("/").}

\item{job_id}{A string specifying the job.}

\item{task_id}{A string specifying the task.}

\item{metadata}{List. Additional metadata to be included in the output. The
paths to the samples, summaries, and model output will be added to the
metadata list.}

\item{diagnostics}{A data.table as returned by \code{\link[=extract_diagnostics]{extract_diagnostics()}}}
}
\value{
Invisible NULL. The function is called for its side effects.
}
\description{
Processes the model fit, extracts samples and quantiles,
and writes them to the appropriate directories.
}
\seealso{
Other write_output: 
\code{\link{sample_processing_functions}},
\code{\link{write_output_dir_structure}()}
}
\concept{write_output}
